requirements.txt
amqp==5.3.1
asgiref==3.9.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
